update test infrastructure + tests for reordering
